+++ lib/src/bones_ui.dart
@@ -1,3 +1,3 @@
 class BonesUI {
-  static const String version = '3.0.9';
+  static const String version = '3.0.10';
 }